requirements.txt
requests
python-dotenv
rich
google-generativeai>=0.8.0
django
djangorestframework
django-cors-headers
